+++ wishbone/lookup/__init__.py
@@ -27,11 +27,3 @@ class Lookup(object):
     pass
 
 from .event import EventLookup
-from .choice import Choice
-from .cycle import Cycle
-from .etcd import ETCD
-from .pid import PID
-from .random_bool import RandomBool
-from .random_integer import RandomInteger
-from .random_word import RandomWord
-from .random_uuid import RandomUUID