+++ wishbone/module/fileout.py
@@ -23,7 +23,7 @@
 #
 
 
-from wishbone import Actor
+from wishbone.actor import Actor
 from wishbone.module import OutputModule
 from wishbone.event import Bulk
 from gevent.os import make_nonblocking