+++ wishbone/module/stdout.py
@@ -22,7 +22,7 @@
 #
 #
 from gevent import monkey; monkey.patch_sys(stdin=False, stdout=True, stderr=False)
-from wishbone import Actor
+from wishbone.actor import Actor
 from wishbone.module import OutputModule
 from os import getpid
 from colorama import init, Fore, Back, Style
@@ -61,7 +61,7 @@ class Format():
         return "PID-%s: %s" % (self.pid_value, data)
 
 
-class STDOUT(Actor, OutputModule):
+class STDOUT(OutputModule):
 
     '''**Prints incoming events to STDOUT.**
 
@@ -132,6 +132,8 @@ class STDOUT(Actor, OutputModule):
             data = "\n".join(data)
         else:
             data = event.get(self.kwargs.selection)
+
+        data = self.encode(data)
 
         output = self.getString(
             getattr(Fore, self.kwargs.foreground_color),